+++ lib/grape/validations/validators/values.rb
@@ -6,7 +6,13 @@ module Grape
           @excepts = options[:except]
           @values = options[:value]
           @proc = options[:proc]
+
+          warn '[DEPRECATION] The values validator except option is deprecated. ' \
+               'Use the except validator instead.' if @excepts
+
           raise ArgumentError, 'proc must be a Proc' if @proc && !@proc.is_a?(Proc)
+          warn '[DEPRECATION] The values validator proc option is deprecated. ' \
+               'The lambda expression can now be assigned directly to values.' if @proc
         else
           @values = options
         end
@@ -17,15 +23,13 @@ module Grape
         return unless params.is_a?(Hash)
         return unless params[attr_name] || required_for_root_scope?
 
-        values = @values.is_a?(Proc) ? @values.call : @values
-        excepts = @excepts.is_a?(Proc) ? @excepts.call : @excepts
         param_array = params[attr_name].nil? ? [nil] : Array.wrap(params[attr_name])
 
         raise Grape::Exceptions::Validation, params: [@scope.full_name(attr_name)], message: except_message \
-          if !excepts.nil? && param_array.any? { |param| excepts.include?(param) }
+          unless check_excepts(param_array)
 
         raise Grape::Exceptions::Validation, params: [@scope.full_name(attr_name)], message: message(:values) \
-          if !values.nil? && !param_array.all? { |param| values.include?(param) }
+          unless check_values(param_array)
 
         raise Grape::Exceptions::Validation, params: [@scope.full_name(attr_name)], message: message(:values) \
           if @proc && !param_array.all? { |param| @proc.call(param) }
@@ -33,9 +37,22 @@ module Grape
 
       private
 
+      def check_values(param_array)
+        values = @values.is_a?(Proc) && @values.arity.zero? ? @values.call : @values
+        return true if values.nil?
+        return param_array.all? { |param| values.call(param) } if values.is_a? Proc
+        param_array.all? { |param| values.include?(param) }
+      end
+
+      def check_excepts(param_array)
+        excepts = @excepts.is_a?(Proc) ? @excepts.call : @excepts
+        return true if excepts.nil?
+        param_array.none? { |param| excepts.include?(param) }
+      end
+
       def except_message
         options = instance_variable_get(:@option)
-        options_key?(:except_message) ? options[:except_message] : message(:except)
+        options_key?(:except_message) ? options[:except_message] : message(:except_values)
       end
 
       def required_for_root_scope?